requirements.txt
pyhamcrest
pytest
allure-pytest